\name{vit-package}
\alias{vit-package}
\docType{package}
\title{
Visual Inference Tools
}
\description{
Traditional methods of teaching statistical
inference rely on concepts that are unfamiliar to students,
such as probability distributions. This may overwhelm a
student's cognitive load and hinder understanding. These
tools simplify the intuition behind inference to a visual
level. They can be used to develop a student's intuitions
about inference before teaching mathematical and
probabilistic concepts.
}
\author{
Simon Potter <spot023@aucklanduni.ac.nz>,
Garrett Grolemund <grolemund@rice.edu>,
Ben Stevenson <bste085@aucklanduni.ac.nz>,
Danny Chang, Vivian Li

Maintainer: Simon Potter <spot023@aucklanduni.ac.nz>
}
\references{
The Visual Inference Tools website: \url{http://www.stat.auckland.ac.nz/~wild/vit/}
}
\keyword{package}
\examples{
\dontrun{iNZightVIT()}
}
