\name{vitBugReport}
\alias{vitBugReport}
\title{
  Writing a Bug Report for \sQuote{VIT}
}
\description{
  This function provides instructions on how to write a bug report that
  is helpful for the \sQuote{VIT} package maintainer.
}
\usage{
vitBugReport()
}
\value{
  Nothing is returned from this function.
}
\author{
  Simon Potter
}
\examples{
  vitBugReport()
}
\keyword{ print }
\keyword{ debugging }
